third_party/WebKit/LayoutTests/http/tests/security/xss-DENIED-getSVGDocument-object.html

<!DOCTYPE html>
<html>
<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
function runTest() {
    test(() => {
      assert_equals(document.getElementById("svgobject").getSVGDocument(), null);
    }, "getSVGDocument() returns null for cross-origin documents");
}
</script>
</head>
<body>
<object onload="runTest()"
        id="svgobject" data="http://localhost:8080/security/resources/empty-svg.php"
        type="image/svg+xml" width="400" height="300"></object> 
</body>
</html>
